requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.8
Django==3.0.5
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
